+++ src/idle-sdk.core/Modules/IModule.cs
@@ -1,0 +1,9 @@
+namespace IdleSdk.Core.Modules;
+
+public interface IModule
+{
+    string Name { get; }
+    Version Version { get; }
+    IReadOnlyCollection<ModuleDependency> Dependencies { get; }
+    void Initialize(ModuleContext context);
+}